pvr: - fix GUIInfoManager CharInfo methods in CPVRManager. fixes crashes and lockups on win32      - don't update timers and recordings when asking for directory contents in GetDirectory() methods.      - sleep a bit longer in the EPG update loop so the GUI can get a full directory listing while an update is running.      - implement compare methods in CEpgInfoTag      - set initial values properly in CPVRManager      - fix bug: timer update was waiting for m_bTriggerRecordingsUpdate to become true      - reduce log spam when updates are requested by addons      - cosmetics: compiler warnings and little typos

+++ xbmc/epg/EpgDatabase.cpp
@@ -362,7 +362,7 @@ int CEpgDatabase::Persist(const CEpg &epg, bool bQueueWrite /* = false */)
   else
   {
     if (ExecuteQuery(strQuery))
-      iReturn = epg.EpgID() <= 0 ? m_pDS->lastinsertid() : epg.EpgID();
+      iReturn = epg.EpgID() <= 0 ? (int) m_pDS->lastinsertid() : epg.EpgID();
   }
 
   return iReturn;
@@ -437,7 +437,7 @@ int CEpgDatabase::Persist(const CEpgInfoTag &tag, bool bSingleUpdate /* = true *
   if (bSingleUpdate)
   {
     if (ExecuteQuery(strQuery))
-      iReturn = m_pDS->lastinsertid();
+      iReturn = (int) m_pDS->lastinsertid();
   }
   else
   {
